[Refactor] clearer variable type announcement

+++ source/core/accelerator/bruteForceAccelerator.cpp
@@ -8,7 +8,7 @@ BruteForceAccelerator::BruteForceAccelerator(std::vector<std::shared_ptr<Interse
 
 bool BruteForceAccelerator::isIntersecting(Ray &ray, Intersection &intersection) {
     bool result = false;
-    for (int i = 0; i < _intersectors.size(); i++) {
+    for (uint64 i = 0; i < _intersectors.size(); i++) {
         result |= _intersectors[i]->isIntersecting(ray, intersection);
     }
 
@@ -17,7 +17,7 @@ bool BruteForceAccelerator::isIntersecting(Ray &ray, Intersection &intersection)
 
 bool BruteForceAccelerator::isOccluded(Ray &ray) {
     bool result = false;
-    for (int i = 0; i < _intersectors.size(); i++) {
+    for (uint64 i = 0; i < _intersectors.size(); i++) {
         result |= _intersectors[i]->isOccluded(ray);
     }
